Add support for staircase light switch
According to the loxone documentation at http://www.loxone.com/enen/service/documentation/api/structure-file.html you have a third command for stairware light switches and this is pulse, this will set the timer and when the time expires the lights will automatically turn of. When using the on command the lights will stay on untill sending the off command.